requirements.txt
Faiss-cpu==1.9.0.post1
Flask==3.1.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
torch==2.5.1
transformers==4.48.1
